requirements.txt
attrs==20.3.0
awscli==1.19.64
awsebcli==3.19.4
bcrypt==3.2.0
blessed==1.18.0
boto3==1.17.62
botocore==1.20.64
cached-property==1.5.2
cement==2.8.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.3
cryptography==3.4.7
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
Flask==1.1.2
future==0.16.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.24
paramiko==2.7.2
pathspec==0.5.9
protobuf==3.15.8
pyasn1==0.4.8
pycparser==2.20
PyMySQL==1.0.2
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.25.1
rsa==4.7.2
s3transfer==0.4.2
semantic-version==2.8.5
six==1.15.0
termcolor==1.1.0
texttable==1.6.3
urllib3==1.25.11
wcwidth==0.1.9
websocket-client==0.58.0
Werkzeug==1.0.1
